nimble_api/requirements.txt
Django==3.2
cffi==1.14.4
djangorestframework==3.11.2
psycopg2-binary==2.8.6
gunicorn==20.0.4
drf-yasg==1.20.0
django-cors-headers==3.5.0
frozendict==1.2
pydash==4.9.0
requests==2.25.1
python-jose==3.2.0
redis==3.5.3
Pillow==8.0.1
python-dateutil==2.8.1
django-ordered-model==3.4.1
django-filter==2.4.0
defusedxml==0.6.0
django-simple-history==2.12.0
django-mptt==0.11.0
django-taggit==1.3.0
django-fsm==2.7.1
requests-oauthlib==1.3.0
packaging==20.9
bcrypt==3.2.0
libgravatar==0.2.4
celery==5.0.2
dj-rest-auth==2.1.2
django-allauth==0.44.0
django-rosetta==0.9.0
django-storages==1.11.1
django-storages[google]==1.11.1
google-auth==1.28.1
sentry-sdk==1.4.3
cryptography==3.1.1
drf-api-tracking==1.7.8
yookassa==2.0.1
django-editorjs-fields==0.2.0
django-nested-inline==0.4.2
django-admin-sortable2==1.0
django-admin-charts==0.24.1
graphene-django==2.15.0
django-renderpdf==3.0.1
WeasyPrint==52.4
djangorestframework-simplejwt==4.6.0
django-mjml==1.0
django-import-export==2.5.0
django-seed==0.2.2
django-admin-interface==0.16.3
whitenoise==5.2.0
boto3==1.25.5
sendgrid==6.9.7
python-dotenv==0.19.2